+++ changelog.md
@@ -1,3 +1,16 @@
+# v0.0.39
+
+## Updates
+
+### Paper
+
+- Added `backgroundColor` prop that accepts `ThemeColors`and overrides the `background-color` of the `Paper`.
+- Added `color` prop that accepts `ThemeColors`and overrides the `color` of the `Paper`.
+
+### Accordion
+
+- Added `hoverColor` prop that accepts `ThemeColors`that will change the hover state color of the default icon.
+
 # v0.0.38
 
 ## Breaking Changes